changes to downbeat distance model (!)
Instead of P0, P is now the feature to observe to set the initial phase. After the first moment,
it deterministically generates new phases and becomes independent of IOIS.

P0 still exists and generates the first value generated by P in the first moment at any moment.

IOI is added for convenience. It calculates the IOI that each B is associated with and is deterministic so model behavior is unaffected.

B-OUT is no longer necessary, since B can use an output formatter.

B is INACTIVE in the first moment. This ALTERS MODEL BEHAVIOR since the pickup is no longer included in the probability of sequences.